PR feedback, make UseXXX methods pub internal
